Moved getNewSourceJob
